requirements.txt
streamlit
python-dotenv 
sentence-transformers
pinecone
pandas
scikit-learn
numpy
